refactor: de-part overture dialogue overlay cluster (#4117)
Convert the five-file overture_dialogue_overlay part cluster to
explicit-import modules with a public state class; shrink the no-part
grandfather allowlist by five entries (234 → 229).

Refs #4117
